+++ SolidworksAddinFramework.Spec/OrientationSpec.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
-using System.Numerics;
+using System.DoubleNumerics;
 using System.Reactive.Disposables;
 using System.Reactive.Linq;
 using System.Reactive.Threading.Tasks;
@@ -26,7 +26,7 @@ namespace SolidworksAddinFramework.Spec
                 var matrix = Matrix4x4.CreateLookAt(Vector3.UnitZ, Vector3.Zero, Vector3.UnitY);
                 view.Orientation3 = math.ToSwMatrix(matrix);
                 await Task.WhenAny(doc.ClearSelectionsNotifyObservable().FirstAsync().ToTask(), Task.Delay(TimeSpan.FromSeconds(30)));
-                matrix = matrix*Matrix4x4.CreateFromAxisAngle(Vector3.UnitY, (float)Math.PI/4);
+                matrix = matrix*Matrix4x4.CreateFromAxisAngle(Vector3.UnitY, (double)Math.PI/4);
                 view.Orientation3 = math.ToSwMatrix(matrix);
                 await Task.WhenAny(doc.ClearSelectionsNotifyObservable().FirstAsync().ToTask(), Task.Delay(TimeSpan.FromSeconds(30)));
             });
