+++ SolidworksAddinFramework.Spec/opengl/Vector3ExtensionsSpec.cs
@@ -1,5 +1,5 @@
 ﻿using System;
-using System.Numerics;
+using System.DoubleNumerics;
 using FluentAssertions;
 using SolidworksAddinFramework.OpenGl;
 using Xunit;
@@ -15,7 +15,7 @@ namespace SolidworksAddinFramework.Spec.opengl
             var r = new Random();
             for (int i = 0; i < 1000; i++)
             {
-                var v = new Vector3((float) r.NextDouble(), (float) r.NextDouble(), (float) r.NextDouble());
+                var v = new Vector3((double) r.NextDouble(), (double) r.NextDouble(), (double) r.NextDouble());
                 var o = v.Orthogonal();
                 v.Dot(o).Should().BeApproximately(0, 1e-9f);
