+++ SolidworksAddinFramework/OpenGl/OpenWire.cs
@@ -2,7 +2,7 @@ using System;
 using System.Collections.Generic;
 using System.Drawing;
 using System.Linq;
-using System.Numerics;
+using System.DoubleNumerics;
 using MathNet.Numerics.LinearAlgebra.Double;
 using OpenTK.Graphics.OpenGL;
 using SolidworksAddinFramework.Geometry;
@@ -14,11 +14,11 @@ namespace SolidworksAddinFramework.OpenGl
     {
         private IReadOnlyList<Vector3> _Points;
         private readonly PrimitiveType _Mode;
-        public float Thickness { get; set; }
+        public double Thickness { get; set; }
 
         public Color Color { get; set; }
 
-        protected Wire(IEnumerable<Vector3> points, float thickness, PrimitiveType mode, Color color)
+        protected Wire(IEnumerable<Vector3> points, double thickness, PrimitiveType mode, Color color)
         {
             _Points = points.ToList();
             Thickness = thickness;
@@ -48,22 +48,22 @@ namespace SolidworksAddinFramework.OpenGl
 
     public class OpenWire : Wire
     {
-        public OpenWire(IEnumerable<Vector3> points, float thickness, Color color)
+        public OpenWire(IEnumerable<Vector3> points, double thickness, Color color)
             : base(points, thickness, PrimitiveType.LineStrip, color)
         { }
 
-        public OpenWire(ICurve curve, double thickness, Color color, double chordTol=1e-6, double lengthTol = 0) : this(curve.GetTessPoints(chordTol, lengthTol), (float)thickness, color)
+        public OpenWire(ICurve curve, double thickness, Color color, double chordTol=1e-6, double lengthTol = 0) : this(curve.GetTessPoints(chordTol, lengthTol), (double)thickness, color)
         {
             
         }
 
-        public OpenWire(IEnumerable<double[]> points, float thickness, Color color) : this(points.Select(p=>p.ToVector3D()), thickness, color)
+        public OpenWire(IEnumerable<double[]> points, double thickness, Color color) : this(points.Select(p=>p.ToVector3D()), thickness, color)
         { }
     }
 
     public class ClosedWire : Wire
     {
-        public ClosedWire(IEnumerable<Vector3> points, float thickness, Color color)
+        public ClosedWire(IEnumerable<Vector3> points, double thickness, Color color)
             : base(points, thickness, PrimitiveType.LineLoop, color)
         { }
     }
